clean up route preview
